tools/llm-extraction/requirements.txt
ruamel.yaml>=0.18.0
pyyaml>=6.0.0
tqdm>=4.66.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
